==> src/datasets/real-world/Dribbble/preprocessing/2. Followers.ipynb <==
import warnings
warnings.filterwarnings("ignore")

import dask.dataframe as dd
import pandas as pd
import json
import sqlite3

from pathlib import Path

# Connect to database sqlite where to store the information regarding the Dribbble dataset.
conn = sqlite3.connect("dribbble_temporary.db")
c = conn.cursor()

# Followers

This file contains information about the **followers** and **followings** of Dribbble's users. 

The corresponding raw data is the file *followers.tsv*.

### Raw data visualization

# Let's try to open the dataframe in order to visualize it.
followers = dd.read_csv(Path("raw_data/followers.tsv"), sep = "\t", header = None, names = ["destination", "info"])

followers.head()

followers.tail()

# Get the number of rows of the dataframe.
followers.shape[0].compute()

Each destination (username) has a list of lists. Each list contains a pair of information: the time (Unix Time) at which the follow action takes place and the following user.

### Unpack information from raw data

# Read the large .tsv file with specified chunksize in order to preprocess it not all together.
chunksize = 10000
followers_chunk = pd.read_csv(Path("raw_data/followers.tsv"), sep = "\t", names = ["destination", "info"], chunksize = chunksize)

chunk_list = list()
users_no_followers = list()

# Each chunk is a dataframe.
for i,chunk in enumerate(followers_chunk):
    # Store information about users with no followers. 
    users_no_followers.append(chunk.loc[(chunk["info"] == "[]")]["destination"].values)
    # Remove empty 'info' rows.
    chunk = chunk.loc[~(chunk["info"] == "[]")]
    if not chunk.empty:
        print("Cumulative not empty chunk: %d" % (chunksize*(i+1)))
        # We evaluate each row that contain strings containing Python expressions.
        chunk["info"] = chunk["info"].apply(lambda x: json.loads(x))
        # Flatten list of list over columns.
        chunk = chunk.explode("info")
        # Extract all the information from this chunk of data.
        chunk = pd.DataFrame(chunk["info"].tolist(), index = chunk["destination"], columns = ["created_at", "source"]).reset_index()

        # Save the current chunk of data into list.
        chunk_list.append(chunk)

# Concatenate all data of the users with at least one follower.
followers = pd.concat(chunk_list, ignore_index = False).reset_index(drop = True)
followers.head()

### Preprocessing

# Load file that maps the multiple usernames of the 'users' table into an unique information.
with open("users_mapper_username.json") as f:
    map_users_multiple_profiles = json.load(f)

# Flatten list of users with no followers.
users_no_followers = pd.Series(users_no_followers).explode().reset_index(drop = True)

# We map the usernames with multiple profiles into 'users' table to have a single value.
# 'destination'.
destination_remap = followers.destination[followers.destination.isin(map_users_multiple_profiles.keys())].map(map_users_multiple_profiles, na_action = "ignore")
print("n° of 'destination' usernames remapped: %d" % len(destination_remap))
followers.destination.loc[destination_remap.index] = destination_remap

# 'source'.
source_remap = followers.source[followers.source.isin(map_users_multiple_profiles.keys())].map(map_users_multiple_profiles, na_action = "ignore")
print("n° of 'source' usernames remapped: %d" % len(source_remap))
followers.source.loc[source_remap.index] = source_remap

# Users with no followers.
users_no_followers_remap = users_no_followers[users_no_followers.isin(map_users_multiple_profiles.keys())].map(map_users_multiple_profiles, na_action = "ignore")
print("n° of usernames with no followers remapped: %d" % len(users_no_followers_remap))
users_no_followers.loc[users_no_followers_remap.index] = users_no_followers_remap

# Check duplicated rows.
followers.duplicated().sum()

# Remove duplicated rows.
followers.drop_duplicates(inplace = True)

# Check duplicated rows.
users_no_followers.duplicated().sum()

# Remove duplicated rows.
users_no_followers.drop_duplicates(inplace = True)

# Check if there are some NaN users ('destination' and 'source').
followers.isna().sum()

# Remove rows with NaN values.
followers.dropna(axis = 0, how = "any", inplace = True)

# Check if there are some NaN users.
users_no_followers.isna().sum()

users_no_followers.dropna(inplace = True)

# Check if there exist an intersection between the users with at least one follower and the users with no followers.
intersection = list(set(followers.destination.unique()) & set(users_no_followers))
intersection

len(intersection)

# Not consider these users into the users with no followers.
users_no_followers = users_no_followers.loc[~users_no_followers.isin(intersection)]

list(set(followers.destination.unique()) & set(users_no_followers))

# Check if the key 'destination' + 'source' is unique.
followers.set_index(["destination", "source"], inplace = True)

followers.index.is_unique

not_unique = followers.index.value_counts()
# The number of couple users with multiple contacts.
print(sum(not_unique > 1))

not_unique = not_unique[not_unique > 1]

not_unique = followers.loc[not_unique.index]
not_unique.head()

followers.reset_index(inplace = True)

# Let's keep the rows less updated regarding 'created_at'.
followers = followers[followers["created_at"] == followers.groupby(["destination", "source"])["created_at"].transform("min")]
followers.head()

followers.set_index(["destination", "source"]).index.is_unique

len(followers)

# Add to the dataframe the information regarding the users with no followers.
no_followers = pd.DataFrame(columns = followers.columns)
no_followers["destination"] = users_no_followers.values

followers = pd.concat([followers, no_followers], ignore_index = False).reset_index(drop = True)

# Convert to datetime UTC.
# We decide to not save the UTC offset (+00:00).
followers.created_at = pd.to_datetime(followers.created_at, unit = "s", utc = True).dt.tz_convert(None)

followers

followers.destination.isna().sum()

# Check if users into 'source' are always contained into 'destination'.
set(followers.source.dropna().unique()).issubset(set(followers.destination.unique()))

# Check users into 'source' that not have any information in 'destination'.
difference = list(set(followers.source.dropna().unique()) - set(followers.destination.unique()))
difference

# Keep only users from 'source' that are also in 'destination'.
followers = followers[(followers.source.isin(followers.destination.unique())) | (followers.source.isna())]

# Check if users into 'source' are always contained into 'destination'.
set(followers.source.dropna().unique()).issubset(set(followers.destination.unique()))

len(followers)

# Check duplicated rows.
followers.duplicated().sum()

followers.destination.isna().sum()

followers.source.isna().sum()

len(followers.destination.unique())

# Save the dataframe into the sql database.
followers.to_sql("followers", conn, index = False, dtype = {"destination": "TEXT", "created_at": "TIMESTAMP", "source": "TEXT"})

conn.close()
