==> src/datasets/real-world/Dribbble/preprocessing/3. Shots.ipynb <==
import warnings
warnings.filterwarnings("ignore")

import dask.dataframe as dd
import pandas as pd
import json
import sqlite3

from pathlib import Path 

pd.set_option("max_colwidth", 30)

# Connect to database sqlite where to store the information regarding the Dribbble dataset.
conn = sqlite3.connect("dribbble_temporary.db")
c = conn.cursor()

# Shots

This file contains information about the **shots** (posts) published by the users on Dribbble. 

The corresponding raw data is the file *shots.tsv*.

### Raw data visualization

# Let's try to open the dataframe in order to visualize it.
shots = dd.read_csv(Path("raw_data/shots.tsv"), sep = "\t", header = None, names = ["author_shot", "info"])

shots.head()

shots.tail()

# Get the number of rows of the dataframe.
shots.shape[0].compute()

Each author has a list of dictionaries. Each dictionary contains the information related to a shot published by the author.

### Unpack information from raw data

# Read the large .tsv file with specified chunksize in order to preprocess it not all together.
chunksize = 10000
shots_chunk = pd.read_csv(Path("raw_data/shots.tsv"), sep = "\t", names = ["author_shot", "info"], chunksize = chunksize)

chunk_list = list()
authors_no_shots = list()

# Each chunk is a dataframe.
for i,chunk in enumerate(shots_chunk):
    # Store information about authors with no shots. 
    authors_no_shots.append(chunk.loc[(chunk["info"] == "[]")]["author_shot"].values)
    # Remove empty 'info' rows.
    chunk = chunk.loc[~(chunk["info"] == "[]")]
    if not chunk.empty:
        print("Cumulative not empty chunk: %d" % (chunksize*(i+1)))
        # We evaluate each row that contain strings containing Python expressions.
        chunk["info"] = chunk["info"].apply(lambda x: json.loads(x))
        # Flatten list of list over columns.
        chunk = chunk.explode("info")
        # Store the information regarding the authors of the shots.
        authors = chunk["author_shot"]
        # Extract all the information from this chunk of data.
        chunk = pd.json_normalize(chunk["info"])
        # Reassign the author feature.
        chunk["author_shot"] = authors.values

        # Save the current chunk of data into list.
        chunk_list.append(chunk)

# Concatenate all data of authors with at least one shot published.
shots = pd.concat(chunk_list, ignore_index = False).reset_index(drop = True)

shots.columns

# Replace '.' character in all column names with '_'.
shots.columns = shots.columns.str.replace(".", "_")

### Preprocessing

# Load file that maps the multiple usernames of the 'users' table into an unique information.
with open("users_mapper_username.json") as f:
    map_users_multiple_profiles = json.load(f)

# Flatten list of authors with no shots.
authors_no_shots = pd.Series(authors_no_shots).explode().reset_index(drop = True)

# We map the usernames with multiple profiles into 'users' table to have a single value.
# 'author'.
author_map = shots.author_shot[shots.author_shot.isin(map_users_multiple_profiles.keys())].map(map_users_multiple_profiles, na_action = "ignore")
print("n° of 'author_shot' usernames remapped: %d" % len(author_map))
shots.author_shot.loc[author_map.index] = author_map

# 'team_username'.
team_username_map = shots.team_username[shots.team_username.isin(map_users_multiple_profiles.keys())].map(map_users_multiple_profiles, na_action = "ignore")
print("n° of 'team_username' usernames remapped: %d" % len(team_username_map))
shots.team_username.loc[team_username_map.index] = team_username_map

# Authors with no shots.
author_map_no_shots = authors_no_shots[authors_no_shots.isin(map_users_multiple_profiles.keys())].map(map_users_multiple_profiles, na_action = "ignore")
print("n° of 'author no shots' usernames remapped: %d" % len(author_map_no_shots))
authors_no_shots.loc[author_map_no_shots.index] = author_map_no_shots

# We can't check the duplicated rows of 'shots' table due to 'tags' feature that contains lists.

# Check if there are some NaN authors into the authors with at least one shot published.
shots.author_shot.isna().sum()

# Check duplicated rows.
authors_no_shots.duplicated().sum()

# Remove duplicated rows.
authors_no_shots.drop_duplicates(inplace = True)

# Check if there are some NaN authors into the authors with no shots.
authors_no_shots.isna().sum()

# Remove NaN values.
authors_no_shots.dropna(inplace = True)

# Check if there exist an intersection between the users with at least one shot and the users with no shots.
intersection = list(set(shots.author_shot.unique()) & set(authors_no_shots))
intersection

len(intersection)

# Not consider these users into the users with no shots.
authors_no_shots = authors_no_shots.loc[~authors_no_shots.isin(intersection)]

list(set(shots.author_shot.unique()) & set(authors_no_shots))

# We delete all the features containing an 'url' information (not of interest).
shots.drop(["html_url", "attachments_url", "buckets_url", "comments_url", "likes_url", "projects_url", "rebounds_url",
            "team_html_url", "team_avatar_url", "team_links_web", "team_links_twitter", "team_buckets_url", 
            "team_followers_url", "team_following_url", "team_likes_url", "team_projects_url", "team_shots_url", 
            "team_members_url", "team_team_shots_url", "rebound_source_url"], axis = 1, inplace = True)

# We also delete the 'images' url (not of interest).
shots.drop(["images_hidpi", "images_normal", "images_teaser"], axis = 1, inplace = True)

# Rename the 'id' column corresponding to the shot.
shots.rename({"id": "shot_id", "tags": "tag"}, axis = 1, inplace = True)

# Add to the dataframe the information regarding the users with no shots.
no_shots = pd.DataFrame(columns = shots.columns)
no_shots["author_shot"] = authors_no_shots.values

shots = pd.concat([shots, no_shots], ignore_index = False).reset_index(drop = True)

shots.head()

# Check if all the dates end with 'Z': Zulu timezone (UTC). Greenwich zone. No time difference between Greenwich Mean Time and Coordinated Universal Time.
print(shots.team_created_at.dropna().str.endswith("Z").all())
print(shots.team_updated_at.dropna().str.endswith("Z").all())
print(shots.created_at.str.endswith("Z").all())
print(shots.updated_at.str.endswith("Z").all())

# Convert to datetime UTC.
# We decide to not save the UTC offset (+00:00).
shots.team_created_at = pd.to_datetime(shots.team_created_at, utc = True).dt.tz_convert(None)
shots.team_updated_at = pd.to_datetime(shots.team_updated_at, utc = True).dt.tz_convert(None)
shots.created_at = pd.to_datetime(shots.created_at, utc = True).dt.tz_convert(None)
shots.updated_at = pd.to_datetime(shots.updated_at, utc = True).dt.tz_convert(None)

# Check if users into 'team_username' are always contained into 'author_shot'.
set(shots.team_username.dropna().unique()).issubset(set(shots.author_shot.unique()))

# There are some team usernames that not have any information about the shots published or not published directly.
difference = list(set(shots.team_username.dropna().unique()) - set(shots.author_shot.unique()))
difference

len(difference)

# Keep only users from 'team_username' that are also in 'author'.
shots = shots[(shots.team_username.isin(shots.author_shot.unique())) | (shots.team_username.isna())]

list(set(shots.team_username.dropna().unique()) - set(shots.author_shot.unique()))

shots.head()

len(shots.shot_id.dropna().unique())

# Check if the 'updated_at' feature is always greater in time than 'created_at' feature.
timediff = shots.updated_at - shots.created_at
sum(timediff < pd.Timedelta(0))

anomalies = shots.loc[timediff < pd.Timedelta(0)]
anomalies[["created_at", "updated_at"]]

# For these anomalies, we decide to swap the 'created_at' and 'updated_at' features.
shots.loc[anomalies.index] = shots.loc[anomalies.index].rename(columns = {"created_at": "updated_at", "updated_at": "created_at"})

timediff = shots.updated_at - shots.created_at
sum(timediff < pd.Timedelta(0))

# Check if the 'team_updated_at' feature is always greater in time than 'team_created_at' feature.
timediff = shots.team_updated_at.dropna() - shots.team_created_at.dropna()
sum(timediff < pd.Timedelta(0))

We want to verify the similarities between the teams defined into *users* table and the teams defined in the current *shots* table.

# Read 'users' table.
users = pd.read_sql("SELECT username, type FROM users", conn)

# Check if the 'team_username' is a subset of teams according to users.
set(shots.team_username.dropna().unique()).issubset(set(users[users.type == "Team"]["username"].unique()))

# Difference. This user is not considered as team into 'users' table but as 'Player'.
list(set(shots.team_username.dropna().unique()) - set(users[users.type == "Team"]["username"].unique()))

len(shots)

# We decide to not consider this user.
shots = shots[(shots.author_shot != "BrodieP") & (shots.team_username != "BrodieP")]

len(shots)

## Tags

We have to extract the tags information that is a list of strings. We decide to create an independent dataframe for this information.

tags = shots[["shot_id", "tag", "author_shot"]].set_index(["shot_id", "author_shot"]).explode("tag").reset_index()
tags

# Check duplicated rows.
tags.duplicated().sum()

# Remove duplicated rows.
tags.drop_duplicates(inplace = True)

len(tags)

len(tags.shot_id.dropna().unique())

len(tags.author_shot.dropna().unique())

# Save the dataframe into the sql database.
tags.to_sql("tags", conn, index = False, dtype = {"shot_id": "INT", "author_shot": "TEXT", "tag": "TEXT"})

## Shots

# Remove 'tags' feature.
shots.drop("tag", axis = 1, inplace = True)

# Check duplicated rows.
shots.duplicated().sum()

# Remove duplicated rows.
shots.drop_duplicates(inplace = True)

len(shots.shot_id.dropna().unique())

shots.author_shot.isna().sum()

len(shots.author_shot.unique())

Let's search the unique key of this table. Let's try the key 'author' + 'team_username' + 'shot_id'.

shots.set_index(["author_shot", "team_username", "shot_id"], inplace = True)

shots.head()

shots.index.is_unique

len(shots.index)

len(shots.index.unique())

# Let's isolate the profiles that not allow to make unique this key.
not_unique = shots.index.value_counts()
print(sum(not_unique > 1))

not_unique = not_unique[not_unique > 1]

unique = shots.drop(not_unique.index, axis = 0)
not_unique = shots.loc[not_unique.index]

not_unique.head()

len(not_unique)

Let's try to reduce to an unique information this subset with multiple profiles keeping the rows more updated regarding the two temporal features 'team_updated_at' and 'updated_at'.

not_unique.reset_index(inplace = True)

# Let's keep the rows more updated regarding 'team_updated_at' (valid for only teams and team members).
update = not_unique.groupby(["author_shot", "team_username", "shot_id"], dropna = False)["team_updated_at"].transform("max")

# Update.
not_unique = not_unique[(not_unique["team_updated_at"] == update) | (update.isna())]

# Let's keep the rows more updated regarding 'updated_at' (valid for all users).
update = not_unique.groupby(["author_shot", "team_username", "shot_id"], dropna = False)["updated_at"].transform("max")

# Update.
not_unique = not_unique[not_unique["updated_at"] == update]

# Check the key.
not_unique.set_index(["author_shot", "team_username", "shot_id"], inplace = True)

not_unique.index.is_unique

len(not_unique.index)

len(not_unique.index.unique())

not_unique.loc[not_unique.index.value_counts() > 1]

# Let's check what are the features (columns) that change for these profiles.
def features_not_unique(x):
    return x.apply(lambda x: x.nunique()-1 if not x.isnull().all() else 0)

# Percentage of change.
not_unique.loc[not_unique.index.value_counts() > 1].reset_index().groupby(["author_shot", "team_username", "shot_id"], dropna = False).apply(features_not_unique).mean(axis = 0).round(3)

not_unique.reset_index(inplace = True)

# Select the rows with more views of the shot.
not_unique = not_unique[not_unique["views_count"] == not_unique.groupby(["author_shot", "team_username", "shot_id"], dropna = False)["views_count"].transform("max")]

not_unique.set_index(["author_shot", "team_username", "shot_id"]).index.is_unique

# Select the rows with more likes of the shot.
not_unique = not_unique[not_unique["likes_count"] == not_unique.groupby(["author_shot", "team_username", "shot_id"], dropna = False)["likes_count"].transform("max")]

not_unique.set_index(["author_shot", "team_username", "shot_id"]).index.is_unique

shots = pd.concat([unique.reset_index(), not_unique], ignore_index = False).reset_index(drop = True)

len(shots)

shots.set_index(["author_shot", "team_username", "shot_id"]).index.is_unique

shots.head()

# Check duplicated rows.
shots.duplicated().sum()

len(shots.shot_id.dropna().unique())

len(shots.author_shot.unique())

shots.set_index(["shot_id", "author_shot"]).index.is_unique

shots.drop("team", axis = 1, inplace = True)

shots.columns

# Save the type for the sqlite table.
dtype = {"shot_id": "INT", "title": "TEXT", "description": "TEXT", "width": "INT", "height": "INT", "views_count": "INT",
         "likes_count": "INT", "comments_count": "INT", "attachments_count": "INT", "rebounds_count": "INT", "buckets_count": "INT",  
         "created_at": "TIMESTAMP", "updated_at": "TIMESTAMP", "animated": "BOOLEAN", "team_username": "TEXT", "author_shot": "TEXT", 
         "team_id": "INT", "team_name": "TEXT", "team_bio": "TEXT", "team_location": "TEXT", "team_buckets_count": "INT", 
         "team_comments_received_count": "INT", "team_followers_count": "INT", "team_followings_count": "INT", 
         "team_likes_count": "INT", "team_likes_received_count": "INT", "team_projects_count": "INT",
         "team_rebounds_received_count": "INT", "team_shots_count": "INT", "team_can_upload_shot": "BOOLEAN", 
         "team_type": "TEXT", "team_pro": "BOOLEAN", "team_created_at": "TIMESTAMP", "team_updated_at": "TIMESTAMP", 
         "team_members_count": "INT"}

# Save the dataframe into the sql database.
shots.to_sql("shots", conn, index = False, dtype = dtype)

conn.close()

### Brief analysis

We analyze the cumulative *_count* features checking for their positive or negative (anomalies) values.

features = ["views_count", "likes_count", "comments_count", "attachments_count", "rebounds_count", 
            "buckets_count"]

for feature in features:
    negatives = len(shots[feature][shots[feature] < 0])
    print("Negative values for '%s': %d of %d" % (feature, negatives, len(shots)))

# Get the subset dataframe with at least one anomaly.
anomalies = shots.loc[(shots[features] < 0).any(axis = 1)]
anomalies.head()

len(anomalies)
