==> src/datasets/real-world/Dribbble/preprocessing/4. Skills.ipynb <==
import warnings
warnings.filterwarnings("ignore")

import pandas as pd
import ast
import sqlite3
import json

from pathlib import Path 

# Connect to database sqlite where to store the information regarding the Dribbble dataset.
conn = sqlite3.connect("dribbble_temporary.db")
c = conn.cursor()

# Skills

This file contains information about the **skills** of the Dribbble's users.

The corresponding raw data is the file *skills.txt*.

N.B. The file regarding this information has been obtained by Francesco.

### Raw data visualization

# Let's try to open the dataframe in order to visualize it.
skills = pd.read_csv(Path("raw_data/skills.txt"), sep = "\t", header = None, names = ["username", "info"])
skills

### Unpack information from raw data

skills["info"] = skills["info"].apply(lambda x: x.rstrip().lstrip())
skills["username"] = skills["username"].apply(lambda x: x.rstrip().lstrip())

# Unique users with ascertain no skills.
users_no_skills = pd.Series(skills[skills["info"] == "[]"].username.unique())

users_no_skills

len(users_no_skills)

Each username has a list containing the information about the skills of the corresponding users. Unfortunately, the *temporal information* on when these skills were earned by users is not present in the data.

# Read the .tsv file with specified chunksize in order to preprocess it not all together.
chunksize = 10000
skills_chunk = pd.read_csv(Path("raw_data/skills.txt"), sep = "\t", names = ["username", "info"], chunksize = chunksize)

chunk_list = list()

# Each chunk is a dataframe.
for i,chunk in enumerate(skills_chunk):
    chunk["info"] = chunk["info"].apply(lambda x: x.rstrip().lstrip())
    chunk["username"] = chunk["username"].apply(lambda x: x.rstrip().lstrip())
    # Remove empty 'info' rows.
    chunk = chunk.loc[~(chunk["info"] == "[]")]
    if not chunk.empty:
        print("Cumulative not empty chunk: %d" % (chunksize*(i+1)))
        # We evaluate each row that contain strings containing Python expressions.
        chunk["info"] = chunk["info"].apply(lambda x: ast.literal_eval(x))
        # Flatten list of list over columns.
        chunk = chunk.explode("info")
        # Rename 'info' column.
        chunk.rename({"info": "skill"}, axis = 1, inplace = True)

        # Save the current chunk of data into list.
        chunk_list.append(chunk)

# Concatenate all data of the users with at least one skill.
skills = pd.concat(chunk_list, ignore_index = False).reset_index(drop = True)
skills

### Preprocessing

# Load file that maps the multiple usernames of the 'users' table into an unique information.
with open("users_mapper_username.json") as f:
    map_users_multiple_profiles = json.load(f)

# We map the usernames with multiple profiles into 'users' table to have a single value.
# 'username'.
username_remap = skills.username[skills.username.isin(map_users_multiple_profiles.keys())].map(map_users_multiple_profiles, na_action = "ignore")
print("n° of 'username' usernames remapped: %d" % len(username_remap))
skills.username.loc[username_remap.index] = username_remap

# Users with no skills.
username_no_skills_remap = users_no_skills[users_no_skills.isin(map_users_multiple_profiles.keys())].map(map_users_multiple_profiles, na_action = "ignore")
print("n° of usernames with no skills remapped: %d" % len(username_no_skills_remap))
users_no_skills.loc[username_no_skills_remap.index] = username_no_skills_remap

# Check duplicated rows.
skills.duplicated().sum()

# Remove duplicated rows.
skills.drop_duplicates(inplace = True)

# Check duplicated rows.
users_no_skills.duplicated().sum()

# Remove duplicated rows.
users_no_skills.drop_duplicates(inplace = True)

# Check NaN values.
skills.isna().sum()

# Check NaN values.
users_no_skills.isna().sum()

# Check if there exist an intersection between the users with at least one skill and the users with no skills.
intersection = list(set(skills.username.unique()) & set(users_no_skills))
intersection

len(intersection)

# Not consider these users into the users with no skills.
users_no_skills = users_no_skills.loc[~users_no_skills.isin(intersection)]

# Add to the dataframe the information regarding the users with no skills.
no_skills = pd.DataFrame(columns = skills.columns)
no_skills["username"] = users_no_skills.values
no_skills["username"] = no_skills["username"].apply(lambda x: x.rstrip().lstrip())

skills = pd.concat([skills, no_skills], ignore_index = False).reset_index(drop = True)

skills

len(skills.username.unique())

# Recheck duplicated rows after this remapping.
skills.duplicated().sum()

# Save the dataframe into the sql database.
skills.to_sql("skills", conn, index = False, dtype = {"username": "TEXT", "skill": "TEXT"})

conn.close()
