==> src/datasets/real-world/Dribbble/preprocessing/5. Comments.ipynb <==
import warnings
warnings.filterwarnings("ignore")

import dask.dataframe as dd
import pandas as pd
import json
import sqlite3

from pathlib import Path 

# Connect to database sqlite where to store the information regarding the Dribbble dataset.
conn = sqlite3.connect("dribbble_temporary.db")
c = conn.cursor()

# Comments

This file contains information about the **comments** on the shots (posts) of Dribbble. 

The corresponding raw data is the file *comments.tsv*.

### Raw data visualization

# Let's try to open the dataframe in order to visualize it.
comments = dd.read_csv(Path("raw_data/comments.tsv"), sep = "\t", header = None, names = ["shot_id", "info"])

comments.head()

comments.tail()

# Get the number of rows of the dataframe.
comments.shape[0].compute()

Each shot has a list of lists. Each list contains the information of the ID of the comment, the date of publication of the comment (Unix Time), the date the comment was modified last time (Unix Time), the author of the comment, the content of the comment and the amount of likes received.

### Unpack information from raw data

# Read the large .tsv file with specified chunksize in order to preprocess it not all together.
chunksize = 10000
comments_chunk = pd.read_csv(Path("raw_data/comments.tsv"), sep = "\t", names = ["shot_id", "info"], chunksize = chunksize)

chunk_list = list()
shots_no_comments = list()

# Each chunk is a dataframe.
for i,chunk in enumerate(comments_chunk):
    # Store information about shots with no comments. 
    shots_no_comments.append(chunk.loc[(chunk["info"] == "[]")]["shot_id"].values)
    # Remove empty 'info' rows.
    chunk = chunk.loc[~(chunk["info"] == "[]")]
    if not chunk.empty:
        print("Cumulative not empty chunk: %d" % (chunksize*(i+1)))
        # We evaluate each row that contain strings containing Python expressions.
        chunk["info"] = chunk["info"].apply(lambda x: json.loads(x))
        # Flatten list of list over columns.
        chunk = chunk.explode("info")
        # Extract all the information from this chunk of data.
        chunk = pd.DataFrame(chunk["info"].tolist(), index = chunk["shot_id"], 
                             columns = ["comment_id", "created_at", "updated_at", "author_comment", "comment", "likes_count"]).reset_index()
        
        # Save the current chunk of data into list.
        chunk_list.append(chunk)

# Concatenate all data of shots with at least one comment.
comments = pd.concat(chunk_list, ignore_index = False).reset_index(drop = True)

comments.head()

### Preprocessing

# Load file that maps the multiple usernames of the 'users' table into an unique information.
with open("users_mapper_username.json") as f:
    map_users_multiple_profiles = json.load(f)

# We map the usernames with multiple profiles into 'users' table to have a single value.
# 'author_comment'
author_map = comments.author_comment[comments.author_comment.isin(map_users_multiple_profiles.keys())].map(map_users_multiple_profiles, na_action = "ignore")
print("n° of 'author_comment' usernames remapped: %d" % len(author_map))
comments.author_comment.loc[author_map.index] = author_map

# Check duplicated rows.
comments.duplicated().sum()

# Check NaN values.
comments.isna().sum()

# Flatten list of shot with no comments.
shots_no_comments = pd.Series(shots_no_comments).explode().reset_index(drop = True)

# Check if there are some NaN shots IDs into the shots with no shots.
shots_no_comments.isna().sum()

shots_no_comments.dropna(inplace = True)

shots_no_comments.duplicated().sum()

shots_no_comments = shots_no_comments.astype(int)

# Check if there exist an intersection between the shot IDs with at least one comment and the shot IDs with no comments.
list(set(comments.shot_id.unique()) & set(shots_no_comments))

# Add to the dataframe the information regarding the shot IDs with no comments.
no_comments = pd.DataFrame(columns = comments.columns)
no_comments["shot_id"] = shots_no_comments.values

comments = pd.concat([comments, no_comments], ignore_index = False).reset_index(drop = True)

# Convert to datetime UTC.
# We decide to not save the UTC offset (+00:00).
comments.created_at = pd.to_datetime(comments.created_at, unit = "s", utc = True).dt.tz_convert(None)
comments.updated_at = pd.to_datetime(comments.updated_at, unit = "s", utc = True).dt.tz_convert(None)

# Check if the 'updated_at' feature is always greater in time than 'created_at' feature.
timediff = comments.updated_at - comments.created_at
sum(timediff < pd.Timedelta(0))

comments

len(comments)

# The unique key of the table.
comments.set_index(["shot_id", "comment_id"]).index.is_unique

comments.duplicated().sum()

comments.shot_id.unique()

comments.shot_id.isna().sum()

comments.columns

# Save the type for the sqlite table.
dtype = {"shot_id": "INT", "comment_id": "INT", "created_at": "TIMESTAMP", "updated_at": "TIMESTAMP", "author_comment": "TEXT", 
         "comment": "TEXT", "likes_count": "INT"}

# Save the dataframe into the sql database.
comments.to_sql("comments", conn, index = False, dtype = dtype)

conn.close()
