==> src/datasets/real-world/Dribbble/preprocessing/6. Likes.ipynb <==
import warnings
warnings.filterwarnings("ignore")

import dask.dataframe as dd
import numpy as np
import pandas as pd
import json
import sqlite3

from pathlib import Path 

# Connect to database sqlite where to store the information regarding the Dribbble dataset.
conn = sqlite3.connect("dribbble_temporary.db")
c = conn.cursor()

# Likes

This file contains information about the **likes** related only to the shots (posts). The information regarding the likes on the comments is available in the 'comments' table (it is available only the information regarding the amount of likes received from the comments).

The corresponding raw data is the file *likes.tsv*.

### Raw data visualization

# Let's try to open the dataframe in order to visualize it.
likes = dd.read_csv(Path("raw_data/likes.tsv"), sep = "\t", header = None, names = ["shot_id", "info"])

likes.head()

likes.tail()

# Get the number of rows of the dataframe.
likes.shape[0].compute()

Each shot has a list of lists. Each list contains information about the id of the like, the time (Unix Time) in which the like was left and the username of the user who left the like, respectively.

### Unpack information from raw data

# Create the likes table into database sqlite.
conn.execute("CREATE TABLE likes (shot_id INT, like_id INT, created_at_unix INT, author_like TEXT)");

# Load file that maps the multiple usernames of the 'users' table into an unique information.
with open("users_mapper_username.json") as f:
    map_users_multiple_profiles = json.load(f)

# Read the large .tsv file with specified chunksize in order to preprocess it not all together.
chunksize = 10000
likes_chunk = pd.read_csv(Path("raw_data/likes.tsv"), sep = "\t", names = ["shot_id", "info"], chunksize = chunksize)

shots_no_likes = list()

# Each chunk is a dataframe.
for i,chunk in enumerate(likes_chunk):
    # Store information about shots with no likes. 
    shots_no_likes.append(chunk.loc[(chunk["info"] == "[]")]["shot_id"].values)
    # Remove empty 'info' rows.
    chunk = chunk.loc[~(chunk["info"] == "[]")]
    if not chunk.empty:
        print("Cumulative not empty chunk: %d" % (chunksize*(i+1)))
        # Read parse a string as a list of lists.
        chunk["info"] = chunk["info"].apply(lambda x: json.loads(x))
        # Flatten list of list over columns.
        chunk = chunk.explode("info")
        # We get the all the data information into list of lists (shot_id, like_id, created_at_unix, author_like).
        data_to_insert = chunk.shot_id.map(lambda x: [x]) + chunk["info"]
        data_to_insert = np.array(data_to_insert.tolist(), dtype = object)

        # We map the usernames with multiple profiles into 'users' table to have a single value.
        def map_func(val, dictionary):
            return dictionary[val] if val in dictionary else val 
        vfunc  = np.vectorize(map_func)
        data_to_insert[:, 3] = vfunc(data_to_insert[:, 3], map_users_multiple_profiles)

        # Insert data into sql table recursively.
        c.executemany("INSERT INTO likes (shot_id, like_id, created_at_unix, author_like) VALUES (?, ?, ?, ?)", data_to_insert)
        conn.commit()

# Flatten list of shots with no likes.
shots_no_likes = pd.Series(shots_no_likes).explode().reset_index(drop = True)

### Preprocessing

# Check if there are some NaN.
shots_no_likes.isna().sum()

shots_no_likes.dropna(inplace = True)

shots_no_likes.duplicated().sum()

shots_no_likes = shots_no_likes.astype(int)

len(shots_no_likes)

likes = pd.read_sql("SELECT shot_id FROM likes", conn)

# Check if there exist an intersection between the shot IDs with at least one like and the shot IDs with no likes.
list(set(likes.shot_id.unique()) & set(shots_no_likes))

# Remove duplicated rows (slow statement).
c.execute("""DELETE FROM likes WHERE rowid NOT IN (SELECT max(rowid) FROM likes GROUP BY like_id)""")
conn.commit()

# Set the right datetime format from Unix Time to UTC creating a new column.
c.execute("""ALTER TABLE likes ADD COLUMN created_at TIMESTAMP""")
c.execute("""UPDATE likes SET created_at = created_at_unix""")
c.execute("""UPDATE likes SET created_at = datetime(created_at_unix, 'unixepoch')""")
conn.commit()

# Inverse operation.
c_new.execute("""UPDATE likes SET created_at_unix = strftime('%s', created_at)""")
conn_new.commit()

# Read the large file with specified chunksize in order to simply fast visualize it.
chunksize = 10000
likes_chunk = pd.read_sql("SELECT * FROM likes", conn, chunksize = chunksize)

next(likes_chunk)

# Add to the dataframe the information regarding the shot IDs with no likes.
no_likes = pd.DataFrame(columns = next(likes_chunk).columns)
no_likes["shot_id"] = shots_no_likes

data_to_insert = no_likes.values.tolist()

# Insert data into sql table recursively.
c.executemany("INSERT INTO likes (shot_id, like_id, created_at_unix, author_like, created_at) VALUES (?, ?, ?, ?, ?)", data_to_insert)
conn.commit()

likes_chunk.close()

# Read 'likes' table.
likes = pd.read_sql("SELECT shot_id, like_id FROM likes", conn)
likes.head()

likes.set_index(["shot_id", "like_id"]).index.is_unique

c.execute("VACUUM"); # This command allows to reaggange database on small data size especially if you have carried out deletions.

conn.close()
