==> src/datasets/real-world/Dribbble/preprocessing/7. Common 'username' and 'shot_id'.ipynb <==
import pandas as pd
import sqlite3

# Connect to database sqlite where the information regarding the Dribbble dataset are stored.
conn = sqlite3.connect("dribbble_temporary.db", detect_types = sqlite3.PARSE_DECLTYPES) # https://docs.python.org/3/library/sqlite3.html
c = conn.cursor()

# Dribbble database (data cleaning)

This notebook deals with fixing some inconsistencies in the Dribbble database. In particular, we notice how the information regarding the users (*username*) and the shots (*shot_id*) are discrepant between the various tables. 

For example, a generic user 'X' could be present in the *users* table but not in the *shots* table where the username field is the identifier (n the *shots* table are stored the information of all the users: both those who have published and those who have not published any shots). This type of anomaly involves also the *shot_id* information among the tables where the *shot_id* field is an identifier. 

Let's find and fix them.

## Intersection between the *username*

In order to solve the anomalies regarding the *username*, we decide to consider as our Dribbble population only the users in common between the tables where the *username* field is an identifier (*users*, *followers*, *skills*, *shots*). 

# Table 'users'.
users = pd.read_sql("SELECT username FROM users", conn)

users.username.isna().sum()

len(users.username.unique())

# Table 'followers'.
followers = pd.read_sql("SELECT destination FROM followers", conn)

followers.destination.isna().sum()

len(followers.destination.unique())

# Table 'skills'.
skills = pd.read_sql("SELECT username FROM skills", conn)

skills.username.isna().sum()

len(skills.username.unique())

# Table 'shots'.
shots = pd.read_sql("SELECT author_shot FROM shots", conn)

shots.author_shot.isna().sum()

len(shots.author_shot.unique())

### Common *username*

intersection_users = list(set.intersection(set(users.username.unique()), set(followers.destination.unique()), set(shots.author_shot.unique()), set(skills.username.unique())))

len(intersection_users)

### Filtering *username*

We save these users into a new table that we will use to filter the usernames into the other tables. 

N.B. We cannot directly filter using a python list because there exist a limit (`SQLITE_LIMIT_VARIABLE_NUMBER`) in the number of values we can provide to execute which is by default set to 999 `?`.

intersection_users = pd.DataFrame(intersection_users, columns = ["username"])

# Save the dataframe into the sql database.
intersection_users.to_sql("intersection_users", conn, index = False, dtype = {"username": "TEXT"})

c.execute("DELETE FROM users WHERE username NOT IN (SELECT username FROM intersection_users)")
conn.commit()

c.execute("DELETE FROM followers WHERE destination NOT IN (SELECT username FROM intersection_users)")
conn.commit()

c.execute("DELETE FROM followers WHERE source NOT IN (SELECT username FROM intersection_users)")
conn.commit()

c.execute("DELETE FROM shots WHERE author_shot NOT IN (SELECT username FROM intersection_users)")
conn.commit()

c.execute("DELETE FROM shots WHERE team_username NOT IN (SELECT username FROM intersection_users)")
conn.commit()

c.execute("DELETE FROM tags WHERE author_shot NOT IN (SELECT username FROM intersection_users)")
conn.commit()

c.execute("DELETE FROM skills WHERE username NOT IN (SELECT username FROM intersection_users)")
conn.commit()

c.execute("DELETE FROM comments WHERE author_comment NOT IN (SELECT username FROM intersection_users)")
conn.commit()

c.execute("DELETE FROM likes WHERE author_like NOT IN (SELECT username FROM intersection_users)")
conn.commit()

These queries allow to filter the usernames keeping the rows where the usernames have `NULL` values.

## Intersection between the *shot_id*

In order to solve the anomalies regarding the *shot_id*, we decide to consider as our Dribbble posts only the shots in common between the tables where the *shot_id* field is an identifier (*shots*, *comments*, *likes*). 

# Table 'shots'.
shots = pd.read_sql("SELECT shot_id FROM shots", conn)

len(shots.shot_id.unique())

# Table 'comments'.
comments = pd.read_sql("SELECT shot_id FROM comments", conn)

comments.shot_id.isna().sum()

len(comments.shot_id.unique())

# Table 'likes'.
likes = pd.read_sql("SELECT shot_id FROM likes", conn)

likes.shot_id.isna().sum()

len(likes.shot_id.unique())

### Common *shot_id*

intersection_shots = list(set.intersection(set(shots.shot_id.dropna().unique()), set(comments.shot_id.unique()), set(likes.shot_id.unique())))

len(intersection_shots)

### Filtering *shot_id*

We save these shots into a new table that we will use to filter the shots into the other tables. 

N.B. We cannot directly filter using a python list because there exist a limit (`SQLITE_LIMIT_VARIABLE_NUMBER`) in the number of values we can provide to execute which is by default set to 999 `?`.

intersection_shots = pd.DataFrame(intersection_shots, columns = ["shot_id"])

# Save the dataframe into the sql database.
intersection_shots.to_sql("intersection_shots", conn, index = False, dtype = {"shot_id": "INT"})

c.execute("DELETE FROM shots WHERE shot_id NOT IN (SELECT shot_id FROM intersection_shots)")
conn.commit()

c.execute("DELETE FROM comments WHERE shot_id NOT IN (SELECT shot_id FROM intersection_shots)")
conn.commit()

c.execute("DELETE FROM likes WHERE shot_id NOT IN (SELECT shot_id FROM intersection_shots)")
conn.commit()

c.execute("DELETE FROM tags WHERE shot_id NOT IN (SELECT shot_id FROM intersection_shots)")
conn.commit()

These queries allow to filter the shot_id keeping the rows when the shot_id have `NULL` values.

# Adjust tables after filtering

Some tables must be adjusted in order to create a database with a correct structure: we delete the rows with `NULL` values in all the columns except one. For example, for the *followers* table:

| destination | created_at | source |
| --- | --- | --- |
| John | NULL | NULL |


The information regarding these cases will be available directly from the features *_count* in the suitable tables. For example, `John` will have 0 `followers_count` in the *users* table. We will build this information in the next notebooks.

## Table *shots*

# Delete NULL rows from the 'shots' table.
c.execute("DELETE FROM shots WHERE shot_id IS NULL")
conn.commit()

## Table *skills*

# Delete NULL rows from the 'skills' table.
c.execute("DELETE FROM skills WHERE skill IS NULL")
conn.commit()

## Table *followers*

# Delete NULL rows from the 'followers' table.
c.execute("DELETE FROM followers WHERE source IS NULL")
conn.commit()

## Table *tags*

# Delete NULL rows from the 'tags' table.
c.execute("DELETE FROM tags WHERE tag IS NULL")
conn.commit()

## Table *comments*

# Delete NULL rows from the 'comments' table.
c.execute("DELETE FROM comments WHERE comment_id IS NULL")
conn.commit()

## Table *likes*

# Delete NULL rows from the 'likes' table.
c.execute("DELETE FROM likes WHERE like_id IS NULL")
conn.commit()

# Remove tables used for intersections.
c.execute("""DROP TABLE intersection_users""")
c.execute("""DROP TABLE intersection_shots""")
conn.commit()

# Release effectively memory from the database.
c.execute("VACUUM");

conn.close()
