==> src/datasets/real-world/Dribbble/preprocessing/10. Teams.ipynb <==
import warnings
warnings.filterwarnings("ignore")

import pandas as pd
import sqlite3

# Connect to database sqlite where is stored the information regarding the Dribbble dataset.
conn = sqlite3.connect("dribbble_temporary.db", detect_types = sqlite3.PARSE_DECLTYPES)
c = conn.cursor()

# Teams

We want to create a table containing the information regarding the **teams**: we keep the rows where the information regarding the 'team_username' is available (not NaN) from 'shots' table. In this manner, we can find the members of the teams (this information is not available from 'users' table). 

Building team structures using this procedure assumes that if a user really belonging to a team has never published a shot, it will not be included in the 'teams' table that we will build below. These users could be into the list of authors with no shots published but we cannot know which team they belong to.

# Read the dataset using pandas.
teams = pd.read_sql("SELECT author_shot, team_username, shot_id, created_at, updated_at, team_id, team_name, team_bio, team_location, team_buckets_count, team_comments_received_count, team_followers_count, team_followings_count, team_likes_count, team_likes_received_count, team_projects_count, team_rebounds_received_count, team_shots_count, team_can_upload_shot, team_type, team_pro, team_created_at, team_updated_at, team_members_count, id_author_shot, id_team_username FROM shots", conn)
teams.head()

teams = teams.loc[teams.team_username.dropna().index]

# Check duplicated rows.
teams.duplicated().sum()

# Rename some column names.
teams.rename({"created_at": "shot_publication_ts", 
              "updated_at": "shot_last_modification_ts", 
              "author_shot": "member_username", 
              "id_author_shot": "id_member_username"}, axis = 1, inplace = True)

teams.set_index(["member_username", "shot_id", "team_username"]).index.is_unique

teams.set_index(["member_username", "shot_id"]).index.is_unique

def add_features(x):
    # Create new features regarding the pubblications of members of team.
    new_features = pd.Series({"n_of_shots": len(x), 
                              "first_shot": min(x.shot_publication_ts), 
                              "last_shot": max(x.shot_last_modification_ts), 
                              "id_member_username": x.id_member_username.unique()[0], 
                              "id_team_username": x.id_team_username.unique()[0]})
    return new_features

teams_feature = teams.groupby(["member_username", "team_username"]).apply(add_features).reset_index()
teams_feature.head()

dtype = {"member_username": "TEXT", "team_username": "TEXT", "n_of_shots": "INT", "first_shot": "TIMESTAMP", "last_shot": "TIMESTAMP", "id_member_username": "INT", "id_team_username": "INT"}

# Save the dataframe into the sql database.
teams_feature.to_sql("teams", conn, index = False, dtype = dtype)

conn.close()
