==> src/datasets/real-world/Dribbble/preprocessing/13. Shots inspection - likes.ipynb <==
import warnings
warnings.filterwarnings("ignore")

import pandas as pd
import numpy as np
import sqlite3
import matplotlib.pyplot as plt
from bisect import bisect
from itertools import chain

# Connect database sqlite where the information regarding the Dribbble dataset are stored.
conn = sqlite3.connect("dribbble.db", detect_types = sqlite3.PARSE_DECLTYPES)
c = conn.cursor()

# Read 'likes' table.
likes = pd.read_sql("SELECT shot_id, like_id, created_at_unix AS created_at_like FROM likes", conn)

# Read 'shots' table.
shots = pd.read_sql("SELECT shot_id, created_at AS created_at_shot, id_author_shot, likes_count AS n_likes_at_end, id_team_username FROM shots", conn)

# Likes information at different temporal horizons from shot publication

We now compute the number of likes occured at different temporal horizons from the corresponding shot publication.

# Create a dataframe where for each 'shot_id' corresponds a list of all the likes creation (ordered in time).
l = likes.groupby("shot_id", as_index = False).agg(created_at_likes = ("created_at_like", lambda x: x.tolist()))
l

shots_likes = pd.merge(shots[["shot_id", "created_at_shot", "id_author_shot"]], l, on = "shot_id", how = "left")

shots_likes.isna().sum()
# There are some shots with no likes.

def likes_after_x_from_shot_publication(df1, df2, x):
    label = f"n_likes_at_{x.replace(" ", "_")}"
    # Convert to seconds.
    seconds = int(pd.to_timedelta(x).total_seconds())
    df1["created_at_shot"] = df1["created_at_shot"] + seconds
    
    mask = df1["created_at_likes"].notnull()
    df1.loc[mask, label] = df1[mask].apply(lambda x: bisect(x["created_at_likes"], x["created_at_shot"]), axis = 1)
    df1[label] = df1[label].fillna(0).astype(int)

    df2 = pd.merge(df2, df1[["shot_id", label]], on = "shot_id", how = "left")
    return df2

# Example.
times = ["1 minutes", "10 minutes", "30 minutes", "1 hours", "8 hours", "16 hours", "1 days", "2 days", "3 days", "5 days", "7 days", "12 days", "20 days", "30 days", "90 days"]

# Convert to unix time.
shots_likes["created_at_shot"] = ((shots_likes["created_at_shot"] - pd.Timestamp("1970-01-01")) // pd.Timedelta("1s"))

df = shots[["shot_id"]]

for time in times:
    print(time)
    df = likes_after_x_from_shot_publication(shots_likes.copy(), df, time)

df

df.to_csv("./raw_data/shots_likes.csv")
